bugfixed

+++ qubit/io/celery/types.py
@@ -2,7 +2,7 @@ import celery
 from typing import NamedTuple
 
 PeriodTask = NamedTuple('PeriodTask', [
-    ('period', int),
+    ('period', float),
     ('task', celery.Task),
     ('name', str)
 ])
